backend/Python/requirements.txt
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
jwt==1.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
